requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.0.8
python-dotenv==1.0.1
pydantic==2.9.2
